refactor: extract auth_models.py + _auth_helpers.py from auth.py
Extract models (TokenResponse, RefreshRequest, UpgradeRequest, LoginRequest)
and password constants to auth_models.py (53 lines).

Extract helper functions (_validate_password, _issue_token_pair,
_set_auth_cookie, _record_failed_login) to _auth_helpers.py (145 lines).

auth.py: 734 → 578 lines (-156, -21.3%). 6 routes intact, 55/55 tests pass.
